requirements.txt
asgiref==3.4.1
autopep8==1.5.7
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.1
h11==0.12.0
idna==3.2
pycodestyle==2.7.0
pydantic==1.8.2
PyJWT==2.1.0
python-dotenv==0.19.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
